Standardize Scala Formatting
* Also fix up statistics generation

+++ utils/src/main/scala/com/reactific/riddl/utils/TreeCopyFileVisitor.scala
@@ -29,9 +29,7 @@ case class TreeCopyFileVisitor(log: Logger, source: Path, target: Path)
   ): FileVisitResult = {
     val resolve = target.resolve(source.relativize(file))
     if (!file.getFileName.startsWith(".")) {
-      if (Files.exists(resolve)) {
-        Files.delete(resolve)
-      }
+      if (Files.exists(resolve)) { Files.delete(resolve) }
       Files.copy(file, resolve, StandardCopyOption.REPLACE_EXISTING)
     }
     FileVisitResult.CONTINUE
